requirements.txt
mxnet-cu90mkl==0.12.1
opencv-python
numpy
nltk
tqdm